requirements.txt
anyio==3.6.1
bcrypt==3.2.2
beanie==1.11.6
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
motor==3.0.0
multidict==6.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pymongo==4.1.1
python-dateutil==2.8.2
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing-extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
yarl==1.7.2
zipp==3.8.1
